// src/app/productos/interfaces/car.interface.ts
export interface CarResponse {
    idalimento:              number;
    inventario_idinventario: number;
    mascota:                 string;
    marca:                   string;
    categoria:               string;
    raza:                    string;
    tamanio:                 number;
    cantidad:                number;
    total:                   number;
    img:                     string;
}